Added support for left ctrl and left shift
now the keyboard is pretty much ready to be used as a main keyboard.
Some very important keys  are missing to be able to program, but these
are very easy to  implement.

Also, now that I am using a keyboard without a right shift, I realized
how much I use both of the shift keys.